+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/AADClientRegistrationRepository.java
@@ -54,8 +54,8 @@ public abstract class AADClientRegistrationRepository implements ClientRegistrat
 
     public boolean isClientNeedConsentWhenLogin(ClientRegistration client) {
         return otherClients.contains(client)
-            && properties.getAuthorization().get(client.getClientName()) != null
-            && !properties.getAuthorization().get(client.getClientName()).isOnDemand();
+            && properties.getAuthorizationClients().get(client.getClientName()) != null
+            && !properties.getAuthorizationClients().get(client.getClientName()).isOnDemand();
     }
 
     public boolean isClientNeedConsentWhenLogin(String id) {
@@ -66,5 +66,9 @@ public abstract class AADClientRegistrationRepository implements ClientRegistrat
     public static boolean isDefaultClient(ClientRegistration clientRegistration) {
         return AZURE_CLIENT_REGISTRATION_ID.equals(
             clientRegistration.getClientName());
+    }
+
+    public static boolean isDefaultClient(String clientId) {
+        return AZURE_CLIENT_REGISTRATION_ID.equals(clientId);
     }
 }