+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/aad/AADAuthenticationProperties.java
@@ -3,8 +3,15 @@
 
 package com.azure.spring.autoconfigure.aad;
 
-import com.azure.spring.aad.webapp.AuthorizationProperties;
+import com.azure.spring.aad.webapp.AuthorizationClientProperties;
 import com.nimbusds.jose.jwk.source.RemoteJWKSet;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.DeprecatedConfigurationProperty;
+import org.springframework.util.ClassUtils;
+import org.springframework.validation.annotation.Validated;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -12,13 +19,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
-import javax.annotation.PostConstruct;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.DeprecatedConfigurationProperty;
-import org.springframework.util.ClassUtils;
-import org.springframework.validation.annotation.Validated;
 
 /**
  * Configuration properties for Azure Active Directory Authentication.
@@ -87,7 +87,7 @@ public class AADAuthenticationProperties {
      */
     private String tenantId = "common";
 
-    private String postLogoutRedirectUri = "{baseUrl}";
+    private String postLogoutRedirectUri;
 
     /**
      * If Telemetry events should be published to Azure AD.
@@ -100,11 +100,11 @@ public class AADAuthenticationProperties {
      */
     private Boolean sessionStateless = false;
 
-    private String authorizationServerUri = "https://login.microsoftonline.com/";
+    private String baseUri = "https://login.microsoftonline.com/";
 
     private String graphMembershipUri = "https://graph.microsoft.com/v1.0/me/memberOf";
 
-    private Map<String, AuthorizationProperties> authorization = new HashMap<>();
+    private Map<String, AuthorizationClientProperties> authorizationClients = new HashMap<>();
 
     @DeprecatedConfigurationProperty(
         reason = "Configuration moved to UserGroup class to keep UserGroup properties together",
@@ -141,28 +141,6 @@ public class AADAuthenticationProperties {
                        .map(AADAuthenticationProperties.UserGroupProperties::getAllowedGroups)
                        .map(allowedGroups -> !allowedGroups.isEmpty())
                        .orElse(false);
-    }
-
-    /**
-     * Validates at least one of the user group properties are populated.
-     *
-     * @throws IllegalArgumentException If no allowed-groups is configured when stateful filter is enabled.
-     */
-    @PostConstruct
-    public void validateUserGroupProperties() {
-        // current implementation is not required, this is only used for compatibility with the previous usage
-        if (authorization.size() > 0 || isResourceServer()) {
-            return;
-        }
-
-        if (this.sessionStateless) {
-            if (allowedGroupsConfigured()) {
-                LOGGER.warn("Group names are not supported if you set 'sessionSateless' to 'true'.");
-            }
-        } else if (!allowedGroupsConfigured()) {
-            throw new IllegalArgumentException("One of the User Group Properties must be populated. "
-                + "Please populate azure.activedirectory.user-group.allowed-groups");
-        }
     }
 
     public boolean isResourceServer() {
@@ -294,12 +272,12 @@ public class AADAuthenticationProperties {
         this.sessionStateless = sessionStateless;
     }
 
-    public String getAuthorizationServerUri() {
-        return authorizationServerUri;
+    public String getBaseUri() {
+        return baseUri;
     }
 
-    public void setAuthorizationServerUri(String authorizationServerUri) {
-        this.authorizationServerUri = authorizationServerUri;
+    public void setBaseUri(String baseUri) {
+        this.baseUri = baseUri;
     }
 
     public String getGraphMembershipUri() {
@@ -310,12 +288,12 @@ public class AADAuthenticationProperties {
         this.graphMembershipUri = graphMembershipUri;
     }
 
-    public Map<String, AuthorizationProperties> getAuthorization() {
-        return authorization;
+    public Map<String, AuthorizationClientProperties> getAuthorizationClients() {
+        return authorizationClients;
     }
 
-    public void setAuthorization(Map<String, AuthorizationProperties> authorization) {
-        this.authorization = authorization;
+    public void setAuthorizationClients(Map<String, AuthorizationClientProperties> authorizationClients) {
+        this.authorizationClients = authorizationClients;
     }
 
     public boolean isAllowedGroup(String group) {