+++ sdk/spring/azure-spring-cloud-context/src/main/java/com/azure/spring/cloud/context/core/impl/ServiceBusTopicSubscriptionManager.java
@@ -3,9 +3,9 @@
 
 package com.azure.spring.cloud.context.core.impl;
 
-import com.microsoft.azure.management.Azure;
-import com.microsoft.azure.management.servicebus.ServiceBusSubscription;
-import com.microsoft.azure.management.servicebus.Topic;
+import com.azure.core.management.exception.ManagementException;
+import com.azure.resourcemanager.servicebus.models.ServiceBusSubscription;
+import com.azure.resourcemanager.servicebus.models.Topic;
 import com.azure.spring.cloud.context.core.config.AzureProperties;
 import com.azure.spring.cloud.context.core.util.Tuple;
 
@@ -14,8 +14,9 @@ import com.azure.spring.cloud.context.core.util.Tuple;
  */
 public class ServiceBusTopicSubscriptionManager extends AzureManager<ServiceBusSubscription, Tuple<Topic, String>> {
 
-    public ServiceBusTopicSubscriptionManager(Azure azure, AzureProperties azureProperties) {
-        super(azure, azureProperties);
+
+    public ServiceBusTopicSubscriptionManager(AzureProperties azureProperties) {
+        super(azureProperties);
     }
 
     @Override
@@ -32,16 +33,20 @@ public class ServiceBusTopicSubscriptionManager extends AzureManager<ServiceBusS
     public ServiceBusSubscription internalGet(Tuple<Topic, String> topicAndSubscriptionName) {
         try {
             return topicAndSubscriptionName.getFirst().subscriptions().getByName(topicAndSubscriptionName.getSecond());
-        } catch (NullPointerException ignore) {
-            // azure management api has no way to determine whether an service bus topic subscription exists
-            // Workaround for this is by catching NPE
-            return null;
+        } catch (ManagementException e) {
+            if (e.getResponse().getStatusCode() == 404) {
+                return null;
+            } else {
+                throw e;
+            }
         }
     }
 
     @Override
     public ServiceBusSubscription internalCreate(Tuple<Topic, String> topicAndSubscriptionName) {
-        return topicAndSubscriptionName.getFirst().subscriptions().define(topicAndSubscriptionName.getSecond())
-            .create();
+        return topicAndSubscriptionName.getFirst()
+                                       .subscriptions()
+                                       .define(topicAndSubscriptionName.getSecond())
+                                       .create();
     }
 }